feat(ci): add pre-commit hook to automatically generate PDFs before committing
Add a `githooks` directory that contains shell scripts, environment files, and Git Hooks to do certain actions automatically.

`export-to-pdf.sh` takes a `.mscz` file and uses MuseScore's CLI to generate a PDF of the score and parts.
`setup-hooks.sh` will copy the contents of `githooks` to `.git/hooks` (which is untracked)